Implemented custom Click Group that ignores group/command callbacks when called with --help
