+++ packages/cli/odahuflow/cli/parsers/deployment.py
@@ -20,15 +20,17 @@ import logging
 import time
 
 import click
+
+from odahuflow.cli.utils import click_utils
 from odahuflow.cli.utils.client import pass_obj
 from odahuflow.cli.utils.error_handler import check_id_or_file_params_present, TIMEOUT_ERROR_MESSAGE
 from odahuflow.cli.utils.output import DEFAULT_OUTPUT_FORMAT, format_output, validate_output_format
 from odahuflow.cli.utils.verifiers import positive_number
 from odahuflow.sdk import config
-from odahuflow.sdk.clients.deployment import ModelDeployment, ModelDeploymentClient, READY_STATE, \
-    FAILED_STATE
 from odahuflow.sdk.clients.api import EntityAlreadyExists, WrongHttpStatusCode
 from odahuflow.sdk.clients.api_aggregated import parse_resources_file_with_one_item
+from odahuflow.sdk.clients.deployment import ModelDeployment, ModelDeploymentClient, READY_STATE, \
+    FAILED_STATE
 
 DEFAULT_WAIT_TIMEOUT = 5
 # 20 minutes
@@ -38,7 +40,7 @@ DEFAULT_DEPLOYMENT_TIMEOUT = 20 * 60
 LOGGER = logging.getLogger(__name__)
 
 
-@click.group()
+@click.group(cls=click_utils.BetterHelpGroup)
 @click.option('--url', help='API server host', default=config.API_URL)
 @click.option('--token', help='API server jwt token', default=config.API_TOKEN)
 @click.pass_context
