+++ packages/cli/odahuflow/cli/parsers/local/packaging.py
@@ -20,6 +20,8 @@ import logging
 from typing import List, Dict, Optional
 
 import click
+
+from odahuflow.cli.utils import click_utils
 from odahuflow.cli.utils.client import pass_obj
 from odahuflow.sdk import config
 from odahuflow.sdk.clients.api_aggregated import \
@@ -33,7 +35,7 @@ from odahuflow.sdk.models import K8sPackager, ModelPackaging, PackagingIntegrati
 LOGGER = logging.getLogger(__name__)
 
 
-@click.group(name='packaging')
+@click.group(name='packaging', cls=click_utils.BetterHelpGroup)
 @click.option('--url', help='API server host', default=config.API_URL)
 @click.option('--token', help='API server jwt token', default=config.API_TOKEN)
 @click.pass_context
