+++ packages/cli/odahuflow/cli/parsers/packaging.py
@@ -22,6 +22,8 @@ from http.client import HTTPException
 
 import click
 from requests import RequestException
+
+from odahuflow.cli.utils import click_utils
 from odahuflow.cli.utils.client import pass_obj
 from odahuflow.cli.utils.error_handler import check_id_or_file_params_present, TIMEOUT_ERROR_MESSAGE
 from odahuflow.cli.utils.logs import print_logs
@@ -39,7 +41,7 @@ LOG_READ_TIMEOUT_SECONDS = 60
 LOGGER = logging.getLogger(__name__)
 
 
-@click.group()
+@click.group(cls=click_utils.BetterHelpGroup)
 @click.option('--url', help='API server host', default=config.API_URL)
 @click.option('--token', help='API server jwt token', default=config.API_TOKEN)
 @click.pass_context
