+++ packages/cli/odahuflow/cli/parsers/packaging_integration.py
@@ -16,6 +16,8 @@
 import http
 
 import click
+
+from odahuflow.cli.utils import click_utils
 from odahuflow.cli.utils.client import pass_obj
 from odahuflow.cli.utils.error_handler import check_id_or_file_params_present
 from odahuflow.cli.utils.output import format_output, DEFAULT_OUTPUT_FORMAT, validate_output_format
@@ -29,7 +31,7 @@ IGNORE_NOT_FOUND_ERROR_MESSAGE = 'Packaging integration {} was not found. Ignore
 ID_AND_FILE_MISSED_ERROR_MESSAGE = 'You should provide a packaging integration ID or file parameter, not both.'
 
 
-@click.group()
+@click.group(cls=click_utils.BetterHelpGroup)
 @click.option('--url', help='API server host', default=config.API_URL)
 @click.option('--token', help='API server jwt token', default=config.API_TOKEN)
 @click.pass_context
