+++ packages/cli/odahuflow/cli/parsers/toolchain_integration.py
@@ -16,6 +16,8 @@
 import http
 
 import click
+
+from odahuflow.cli.utils import click_utils
 from odahuflow.cli.utils.client import pass_obj
 from odahuflow.cli.utils.error_handler import IGNORE_NOT_FOUND_ERROR_MESSAGE, \
     check_id_or_file_params_present
@@ -27,7 +29,7 @@ from odahuflow.sdk.clients.toolchain_integration import ToolchainIntegrationClie
 from odahuflow.sdk.models import ToolchainIntegration
 
 
-@click.group()
+@click.group(cls=click_utils.BetterHelpGroup)
 @click.option('--url', help='API server host', default=config.API_URL)
 @click.option('--token', help='API server jwt token', default=config.API_TOKEN)
 @click.pass_context
